+++ auth_domain/src/main/java/com/example/auth_domain/repository/StudentsRepository.kt
@@ -1,0 +1,23 @@
+package com.example.auth_domain.repository
+
+import com.example.auth_domain.param.RegisterParam
+import com.example.auth_domain.param.ResetPasswordParam
+
+interface StudentsRepository {
+
+    suspend fun register(
+        registerParam: RegisterParam,
+    )
+
+    suspend fun duplicateCheckId(
+        accountId: String
+    )
+
+    suspend fun duplicateCheckEmail(
+        email: String
+    )
+
+    suspend fun resetPassword(
+        resetPasswordParam: ResetPasswordParam
+    )
+}
